♻️ isolate memory permission payloads
